Fix: Case when Time or Momory are Hidden
Apparently there are situations where for a submission the time and memory are hidden, for such cases the crawler would fail, since it expects double as an input and not a string. String parsing is alwasys risky. We need to come up with a better approach, as of now just added a try catch block and crawler won't fetch best submission in such scenerios, it would just return the first AC solution.

Issue Pointed by dhruvsomani:
Discussion Link : https://discuss.codechef.com/questions/90609/a-tool-to-download-all-your-successful-codechef-solutions-in-one-go